Create module Display
